+++ src/vscodiff/common/text_edit.py
@@ -1,10 +1,7 @@
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
 from typing import Callable
 
-from vscodiff.common.asserts import check_adjacent_items
 from vscodiff.common.position import Position
-from vscodiff.common.position_to_offset import PositionOffsetTransformer
 from vscodiff.common.range import Range
 from vscodiff.common.text_length import TextLength
 from vscodiff.common.uint import Constants
@@ -62,39 +59,3 @@ class LineBasedText(AbstractText):
 class ListText(LineBasedText):
     def __init__(self, lines: list[str]):
         super().__init__(lambda n: lines[n - 1], len(lines))
-
-
-class StringText(AbstractText):
-    def __init__(self, value: str):
-        super().__init__()
-
-        self.value = value
-
-        self._t = PositionOffsetTransformer(value)
-
-    @property
-    def length(self):
-        return self._t.text_length
-
-    def get_value_of_range(self, range_: Range):
-        return self._t.get_offset_range(range_).substring(self.value)
-
-
-@dataclass
-class SingleTextEdit:
-    range: Range
-    text: str
-
-    @property
-    def is_empty(self):
-        return self.range.is_empty() and len(self.text) == 0
-
-
-@dataclass(init=False)
-class TextEdit:
-    edits: list[SingleTextEdit]
-
-    def __init__(self, edits: list[SingleTextEdit]):
-        assert check_adjacent_items(edits, lambda a, b: a.range.end <= b.range.start)
-
-        self.edits = edits